Addressed comments and created new shared_constants.h header
