+++ include/xdg/gprt/shared_structs.h
@@ -1,5 +1,5 @@
 #include "gprt.h"
-#include "../shared_enums.h"
+#include "../shared_constants.h"
 #include "../geometry/dp_math.h"
